+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/DockerRegistryAPIHelperIT.java
@@ -17,18 +17,8 @@
 package io.dockstore.client.cli;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.github.dockerjava.api.DockerClient;
-import com.github.dockerjava.api.command.PullImageResultCallback;
-import com.github.dockerjava.api.exception.DockerClientException;
-import com.github.dockerjava.api.model.SearchItem;
-import com.github.dockerjava.core.DefaultDockerClientConfig;
-import com.github.dockerjava.core.DockerClientConfig;
-import com.github.dockerjava.core.DockerClientImpl;
-import com.github.dockerjava.httpclient5.ApacheDockerHttpClient;
-import com.github.dockerjava.transport.DockerHttpClient;
 import io.dockstore.common.CommonTestUtilities;
 import io.dockstore.common.NonConfidentialTest;
 import io.dockstore.common.Registry;
@@ -36,8 +26,6 @@ import io.dockstore.webservice.DockstoreWebserviceApplication;
 import io.dockstore.webservice.DockstoreWebserviceConfiguration;
 import io.dockstore.webservice.helpers.DockerRegistryAPIHelper;
 import io.dropwizard.testing.DropwizardTestSupport;
-import java.io.IOException;
-import java.util.List;
 import java.util.Optional;
 import okhttp3.Response;
 import org.junit.jupiter.api.AfterAll;
@@ -88,33 +76,5 @@ class DockerRegistryAPIHelperIT {
         assertTrue(manifestResponse.get().isSuccessful());
         calculatedDigest = "sha256:" + DockerRegistryAPIHelper.calculateDockerImageDigest(manifestResponse.get());
         assertEquals(digest, calculatedDigest);
-    }
-
-    @Test
-    void testOldSchemaDockerImages() throws IOException, InterruptedException {
-        DockerClientConfig config = DefaultDockerClientConfig.createDefaultConfigBuilder().build();
-
-        try (DockerHttpClient httpClient = new ApacheDockerHttpClient.Builder().dockerHost(config.getDockerHost())
-            .sslConfig(config.getSSLConfig()).build(); DockerClient dockerClient = DockerClientImpl.getInstance(config, httpClient)) {
-
-            // this does not return enough image to diagnose Docker image version
-            List<SearchItem> items = dockerClient.searchImagesCmd("Java").exec();
-
-            dockerClient.pullImageCmd("weischenfeldt/pcawg_sv_merge")
-                .withTag("1.0.2")
-                .exec(new PullImageResultCallback())
-                .awaitCompletion();
-
-            DockerClientException exception = assertThrows(DockerClientException.class, () -> {
-                // note this just dies and you need a Docker client on the localhost to assess whether an image is the older schema which is than ideal
-                // see https://docs.docker.com/engine/deprecated/#pushing-and-pulling-with-image-manifest-v2-schema-1
-                // see also https://github.com/dockstore/dockstore/issues/5878
-                dockerClient.pullImageCmd("memcached")
-                    .withTag("1.4.22")
-                    .exec(new PullImageResultCallback())
-                    .awaitCompletion();
-            });
-            assertTrue((exception.getMessage().contains("Docker Image Format v1 and Docker Image manifest version 2, schema 1 support is disabled")));
-        }
     }
 }
